created separate one click quick play option
