+++ static/js/quick_play.js
@@ -1,68 +1,26 @@
+import { getGeneratedPrompt } from "./modules/generator.js"
+import { articleCheck } from "./modules/wikipediaAPI/util.js";
 
-import { autocompleteInput } from "./modules/autocomplete.js";
-import { PromptGenerator } from "./modules/generator.js"
-import { getArticleTitle, articleCheck } from "/static/js/modules/wikipediaAPI/util.js";
+const difficulty = 1000;
 
 let app = new Vue({
-    delimiters: ['[[', ']]'],
     el: '#app',
-    components: {
-        'prompt-generator': PromptGenerator,
-        'ac-input': autocompleteInput
-    },
 
-    data: {
+    mounted: async function() {
+        let start = '', end = '';
 
-        start: "", // The input article names
-        end: "",
-
-        startPrompt: "",  // The actual wikipedia article names that will be played
-        endPrompt: "", 
-
-        articleCheckMessage: "",
-
-        autocomplete: null
-    },
-
-    methods: {
-
-        async checkArticles() {
-            this.articleCheckMessage = "";
-            
-            if(this.start == "" || this.end == ""){
-                this.articleCheckMessage = "Prompt is currently empty";
-                return;
-            }
-            this.startPrompt = await getArticleTitle(this.start);
-            if (this.startPrompt === null) {
-                this.articleCheckMessage = `"${this.start}" is not a Wikipedia article`;
-                return;
-            }
-
-            this.endPrompt = await getArticleTitle(this.end);
-            if (this.endPrompt === null) {
-                this.articleCheckMessage = `"${this.end}" is not a Wikipedia article`;
-                return;
-            }
-
-            const checkRes = await articleCheck(this.endPrompt);
-            if ('warning' in checkRes) {
-                this.articleCheckMessage = checkRes["warning"];
-                return;
-            }
-        },
-
-        async play() {
-
-            await this.checkArticles();
-            if(this.articleCheckMessage != "") return;
-
-            console.log(`${this.startPrompt} -> ${this.endPrompt}`);
-            
-            const start_param = encodeURIComponent(this.startPrompt);
-            const end_param = encodeURIComponent(this.endPrompt);
-            window.location.replace(`/play/${start_param}/${end_param}`);
+        // loop until there is a valid start and end
+        while(true){
+            [start, end] = await getGeneratedPrompt(difficulty, 2);
+            const checkRes = await articleCheck(end);
+            if (!('warning' in checkRes)) break;
         }
-	}
 
+        console.log(`${start} -> ${end}`);
+            
+        const start_param = encodeURIComponent(start);
+        const end_param = encodeURIComponent(end);
+
+        window.location.replace(`/play/${start_param}/${end_param}`);
+    }
 });
