+++ templates/home.html
@@ -68,7 +68,10 @@
                 <a href="/login">Log in</a>/<a href="/register">Register</a> to create a new private lobby and play with your friends!
             </p>
             <p>
-                Play a Quick Game <a href="/quick_run">here</a>!
+                Play a Quick Game <a href="/quick_play">here</a>!
+            </p>
+            <p>
+                Play your own prompt <a href="/custom_play">here</a>!
             </p>
         </template>
     </div>
